Feat: First draft of whatsapp notifications refactor
